Build 104
Merge branch 'ProBetaLTS' of github.com:Phyronnaz/VoxelPrivate into ProBetaLTS

🐛 More fixes

🐛 Fix VS2017 error

🐛 Fix UE4

♻ Update build script

✨ UE5

+++ Source/VoxelEditor/Private/Details/VoxelPaintMaterialCustomization.cpp
@@ -360,7 +360,7 @@ void FVoxelPaintMaterial_MaterialCollectionChannelCustomization::CustomizeHeader
 
 					Material->SetScalarParameterValue("EditorPreviewSingleIndex", Index);
 
-					const FVoxelMaterialCollectionMaterialInfo MaterialInfo{ static_cast<uint8>(Index), Material, *FString::Printf(TEXT("Index %03d"), Index) };
+					const FVoxelMaterialCollectionMaterialInfo MaterialInfo{ Index, Material, *FString::Printf(TEXT("Index %03d"), Index) };
 					AssetsToMaterials->Add(Material, MaterialInfo);
 					IndicesToMaterials->Add(Index, MaterialInfo);
 				}
